requirements.txt
doit>=0.29.0
pytest>=2.7.2
pytest-cov>=1.8.1
pylint>=1.4.4
ruamel.yaml>=0.15.35
